# perf: speed up clean operation