@emcsween 2022-03-19
Thanks for the review, @manast! I changed the deletion marker to the empty string and reran the tests.