@emcsween 2022-03-20
I added a delay to the test I added in the hope that it will help it pass in CI.