@manast 2022-03-21
:tada: This PR is included in version 4.8.1 :tada:

The release is available on:
- [npm package (@latest dist-tag)](https://www.npmjs.com/package/bull/v/4.8.1)
- [GitHub release](https://github.com/OptimalBits/bull/releases/tag/v4.8.1)

Your **[semantic-release](https://github.com/semantic-release/semantic-release)** bot :package::rocket: